lekki refactor
- podzieliłem pliki na foldery
- wydzieliłem plik przechowujący parametry algorytmów
- wydzieliłem z series_classification.py tworzenie modeli do create_class_models
