+++ src/params.py
@@ -1,0 +1,22 @@
+from geneticalgorithm2 import AlgorithmParams
+
+_GA_PARAMS = {
+    "variable_type": "real",
+    "algorithm_parameters": AlgorithmParams(
+        max_num_iteration=20,
+        population_size=100,
+        max_iteration_without_improv=50,
+        mutation_probability=0.05,
+    ),
+}
+_GA_RUN_PARAMS = {
+    "no_plot": True,
+    "disable_printing": False,
+    "disable_progress_bar": False,
+}
+
+cmeans_params = {
+    "m": 2,
+    "error": 1e-8,
+    "maxiter": 1e2
+}
