+++ src/test_cases.py
@@ -4,7 +4,7 @@ Module with test cases.
 
 import sys
 import os
-import series_classification
+import classification.series_classification as series_classification
 
 
 class TestCase:  # pylint: disable=too-few-public-methods
@@ -45,6 +45,8 @@ class TestCase:  # pylint: disable=too-few-public-methods
         None.
 
         """
+
+        # Train
         series_classifier = series_classification.train(
             os.path.join(dataset_dir, "Train"),
             self._previous_considered_indices,
@@ -53,6 +55,8 @@ class TestCase:  # pylint: disable=too-few-public-methods
         )
         name = os.path.basename(os.path.normpath(dataset_dir))
         series_classifier.save(f"{name}_model{run_id}.dat")
+        
+        # Test
         for length_percent in self._length_percentages:
             train_result = series_classification.test(
                 os.path.join(dataset_dir, "Train"),
@@ -83,10 +87,11 @@ class TestCase:  # pylint: disable=too-few-public-methods
 
 
 if __name__ == "__main__":
-    #if len(sys.argv) < 2:
-    #    raise RuntimeError("Too few arguments, dataset directory path is missing")
-    #data_dir = sys.argv[1]
-    data_dir = "tests/UWaveGestureLibrary_Preprocessed"
+    if len(sys.argv) < 2:
+        data_dir = "..\\UWaveGestureLibrary_Preprocessed"
+        # raise RuntimeError("Too few arguments, dataset directory path is missing")
+    else:
+        data_dir = sys.argv[1]
 
     tests = []
     tests.append(TestCase([1, 0.8], [1, 2, 3, 4], 1, 3))
